pyproject.toml
[project]
name = "vl8"
version = "0.2.0"
description = "🔉 Perturbed audio 🔉"
authors = [{ name = "Tom Ritchford", email = "tom@swirly.com" }]
requires-python = ">=3.10"
readme = "README.md"
license = "MIT"
classifiers = ["Programming Language :: Python :: 3", "Programming Language :: Python :: 3.10", "Programming Language :: Python :: 3.11", "Programming Language :: Python :: 3.12", "Programming Language :: Python :: 3.13", "Programming Language :: Python :: 3.14"]
dependencies = [
    "abbrev>=1.0.1,<2",
    "graycode>=1.0.5,<2",
    "more-itertools>=9.0.0,<10",
    "numpy>=1.24.1,<2",
    "pydub>=0.25.1,<0.26",
    "pyyaml>=6.0,<7",
    "runs>=1.1.0,<2",
    "soundfile>=0.11.0,<0.12",
    "stroll>=1.1.0,<2",
    "tdir>=1.4.1,<2",
    "toml>=0.10.2,<0.11",
    "typeguard>=2.13.3,<3",
    "wavemap>=2.0.1,<3",
    "xmod>=1.3.2,<2",
]

[tool.doks]
auto = true

[tool.ruff]
[dependency-groups]
dev = [
    "coverage>=7.1.0,<8",
    "impall>=1.1.1,<2",
    "pytest>=7.2.1,<8",
    "pyupgrade>=3.21.2",
    "ruff>=0.14.14",
    "tdir>=1.4.1,<2",
    "ty>=0.0.14",
]

[tool.uv]

[tool.uv.build-backend]
module-root = ""

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"
[tool.coverage.run]
branch = true
source = "vl8"

[tool.coverage.report]
fail_under = "60"
skip_covered = true
exclude_lines = ["pragma: no cover", "if False:", "if __name__ == .__main__.:", "raise NotImplementedError"]

